# Conflicts:
#	PVT.Q1.2017.Shop/Shop.Common/Models/Currency.cs

+++ PVT.Q1.2017.Shop/Shop.Common/Models/CurrencyRate.cs
@@ -20,10 +20,16 @@ namespace Shop.Common.Models
     /// </summary>
     public class CurrencyRate : BaseEntity
     {
+        #region Properties
+
         /// <summary>
         ///     Gets or sets the cross course.
         /// </summary>
-        public double CrossCourse { get; set; }
+        public decimal CrossCourse { get; set; }
+
+        #endregion //Properties
+
+        #region Foreign Keys
 
         /// <summary>
         /// Currency id
@@ -35,6 +41,10 @@ namespace Shop.Common.Models
         /// </summary>
         public int TargetCurrencyId { get; set; }
 
+        #endregion //Foreign Keys
+
+        #region Navigation Properties
+
         /// <summary>
         /// Gets or sets the currency.
         /// </summary>
@@ -44,5 +54,7 @@ namespace Shop.Common.Models
         ///     Gets or sets the target currency.
         /// </summary>
         public virtual Currency TargetCurrency { get; set; }
+
+        #endregion //Navigation Properties
     }
 }
